Move ForbiddenGuard to common
